+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## v1.2.0
+
+Add ability to see closed/merged PRs
+Add repository name to tooltip
+
 ## v1.1.2
 
 Fix to only pull user's pull requests